% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcell-package.R
\docType{data}
\name{datascreen}
\alias{datascreen}
\title{Bone mineral density data}
\format{
A data frame with 84 rows and 101 columns.
}
\source{
https://www.ebi.ac.uk/biostudies/files/E-MEXP-1618/Normarrayexpressdata.txt.magetab
 https://www.ebi.ac.uk/biostudies/files/E-MEXP-1618/E-MEXP-1618.sdrf.txt
}
\description{
A subset of data from the European Bioinformatics Institute Array-Express repository.
The BMD data consists of gene expression measurements of 54,675 probes of 84 Norwegian women.
}
\details{
Given the large number of variables in the dataset, a pre-screening step was implemented to identify the subset of variables that are most correlated with the outcome of interest, the total hip T-score.
To accomplish this, we first log-transformed all the predictors and then utilized the robust correlation estimate based on Winsorization.
The screened data comprise measurements of $p = 100$ genes from $n = 84$ Norwegian women.
}
\examples{
data(datascreen)
}
\keyword{datasets}
