% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_regression.R
\name{lambdamax_beta}
\alias{lambdamax_beta}
\title{function to calculate the max lambda}
\usage{
lambdamax_beta(
  y,
  x,
  betahat,
  intercept,
  deltahat,
  zetahat,
  softbeta,
  lambda_delta,
  softdelta,
  lambda_zeta,
  softzeta,
  alpha,
  tol,
  maxiter
)
}
\description{
function to calculate the max lambda
}
