% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{lassocv}
\alias{lassocv}
\title{Lasso with cross validation}
\usage{
lassocv(y, x)
}
\arguments{
\item{y}{response}

\item{x}{the design matrix}
}
\value{
betahat: the estimated regression coefficient vector
}
\description{
Lasso with cross validation
}
\examples{
data = genevar()
y = data$y
x = data$x
fit = lasso.cv(y,x)

}
