% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcellmap.R
\name{newcellmap}
\alias{newcellmap}
\title{cellmap}
\usage{
newcellmap(
  R,
  indcells = NULL,
  indrows = NULL,
  standOD = NULL,
  showVals = NULL,
  D = NULL,
  rowlabels = "",
  columnlabels = "",
  mTitle = "",
  rowtitle = "",
  columntitle = "",
  showrows = NULL,
  showcolumns = NULL,
  nrowsinblock = NULL,
  ncolumnsinblock = NULL,
  manualrowblocksizes = NULL,
  manualcolumnblocksizes = NULL,
  rowblocklabels = NULL,
  columnblocklabels = NULL,
  autolabel = TRUE,
  columnangle = 90,
  sizemain = 2,
  sizetitles = 1.1,
  adjustrowlabels = 1,
  adjustcolumnlabels = 1,
  colContrast = 1,
  outlyingGrad = TRUE,
  darkestColor = sqrt(qchisq(0.999, 1)),
  drawCircles = FALSE
)
}
\description{
from cellWise::cellMap, only changed the background colour
}
