% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcell-package.R
\name{regcell}
\alias{regcell}
\title{Cellwise regularized robust sparse regression}
\description{
Cellwise regularized robust sparse regression
}
\references{
Su P, Tarr G, Muller S and Wang S (2023). CR-Lasso: Robust cellwise regularized sparse regression. arXiv:2307.05234
}
\author{
Peng Su (peng.su@sydney.edu.au), Garth Tarr (garth.tarr@gmail.com), Samuel Muller (samuel.muller@mq.edu.au) and Suojin Wang (sjwang@stat.tamu.edu)
}
\keyword{package}
