% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{Rlars}
\alias{Rlars}
\title{robust Lars}
\usage{
Rlars(y, x)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}
}
\value{
betahat: estimated beta

sigmahat: estimated scale

res: residuals
}
\description{
robust Lars
}
\examples{

data = genevar()
y = data$y
x = data$x
fit = Rlars(y,x)
}
