% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{SLTS}
\alias{SLTS}
\title{Sparse LTS with BIC}
\usage{
SLTS(y, x)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}
}
\value{
betahat: estimated beta
}
\description{
Sparse LTS with BIC
}
\examples{

data = genevar()
y = data$y
x = data$x
fit = Rlars(y,x)
}
