% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_regression.R
\name{sregcell}
\alias{sregcell}
\title{robust sparse regression under cellwise contamination (with a grid of lambdas)}
\usage{
sregcell(
  y,
  x,
  betahat = NULL,
  intercept = NULL,
  softbeta = TRUE,
  softdelta = TRUE,
  softzeta = TRUE,
  lambda_delta = 2.56,
  lambda_zeta = 2.56,
  alpha = 0.5,
  penal = 1,
  penaldelta = 1,
  tol = 0.001,
  maxiter = 100
)
}
\arguments{
\item{y}{n-dimensional response vector}

\item{x}{nxp design matrix}

\item{betahat}{p-dimensional initial estimate of beta}

\item{intercept}{initial estimate of intercept}

\item{softbeta}{whether to use soft/hard threshold for beta}

\item{softdelta}{whether to use soft/hard threshold for delta (outliers in x)}

\item{softzeta}{whether to use soft/hard threshold for zeta (outliers in y)}

\item{lambda_delta}{tuning parameter of delta}

\item{lambda_zeta}{tuning parameter of zeta}

\item{alpha}{the importance factor of the regression loss (between 0-1, by default is 0.5)}

\item{penal}{the penalty parameter for model selection (by default is 1, equivalent to BIC )}

\item{penaldelta}{the penalty of number of detected outliers (for debug, by default is 0)}

\item{tol}{the tolerance of convergence, by default is 1e-3}

\item{maxiter}{number of iterations, by default is 100}
}
\value{
betahat: the estimated beta

intercept_hat: the estiamted intercept

betahat_opt: the estimated beta with post-cellwise-robust regression

intercept_opt: the estimated intercept with post-cellwise-robust regression

allfits: a list of results including all outputs from all candidate tuning parameters

...
}
\description{
robust sparse regression under cellwise contamination (with a grid of lambdas)
}
\examples{

data = genevar()
y = data$y
x = data$x
fit = sregcell(y,x)

}
