% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_regression.R
\name{sregcell_lambda}
\alias{sregcell_lambda}
\title{Cellwise regularized robust sparse regression (with a specific lambda)}
\usage{
sregcell_lambda(
  y,
  x,
  softbeta = TRUE,
  softdelta = TRUE,
  softzeta = TRUE,
  lambda_delta = 2.56,
  lambda_zeta = 2.56,
  lambda = 0,
  alpha = 0.5,
  tol = 0.001,
  maxiter = 100
)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}

\item{softbeta}{whether to use soft/hard threshold for beta}

\item{softdelta}{whether to use soft/hard threshold for delta(outliers in x)}

\item{softzeta}{whether to use soft/hard threshold for zeta(outliers in y)}

\item{lambda_delta}{tuning parameter of delta}

\item{lambda_zeta}{tuning parameter of zeta}

\item{lambda}{tuning parameter of beta}

\item{alpha}{the importance factor of the regression loss (between 0-1, by default is 0.5)}

\item{tol}{the tolerence of convergence}

\item{maxiter}{the number of interations}
}
\value{
fit
}
\description{
Cellwise regularized robust sparse regression (with a specific lambda)
}
\examples{


data = genevar()
y = data$y
x = data$x
fit = sregcell_lambda(y,x, lambda = 1)

}
