% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_regression.R
\name{sregcell_std}
\alias{sregcell_std}
\title{robust sparse regression under cellwise contamination with a grid of lambdas (standardize predictors first)}
\usage{
sregcell_std(
  y,
  x,
  scale.method = qnscale,
  df = Inf,
  softbeta = TRUE,
  softdelta = TRUE,
  softzeta = TRUE,
  lambda_delta = NULL,
  lambda_zeta = 1,
  prob = 0.995,
  alpha = 0.5,
  penal = 1,
  penaldelta = 0,
  tol = 0.001,
  maxiter = 30
)
}
\arguments{
\item{y}{n-dimensional response vector}

\item{x}{nxp design matrix}

\item{scale.method}{method we used to obtain robust scales, by default is qn}

\item{df}{degrees of freedom of the assuming distribution of predictors}

\item{softbeta}{whether to use soft/hard threshold for beta}

\item{softdelta}{whether to use soft/hard threshold for delta(outliers in x)}

\item{softzeta}{whether to use soft/hard threshold for zeta(outliers in y)}

\item{lambda_delta}{tuning parameter of delta}

\item{lambda_zeta}{tuning parameter of zeta}

\item{prob}{probability of quantiles, by default is 0.995}

\item{alpha}{the importance factor of the regression loss (between 0-1, by default is 0.5)}

\item{penal}{the penalty parameter for model selection (by default is 1, equivalent to BIC )}

\item{penaldelta}{the penalty of number of detected outliers (for further development, by default is 0)}

\item{tol}{the tolerance of convergence, by default is 1e-3}

\item{maxiter}{number of iterations, by default is 100}
}
\value{
betahat: the estimated beta

intercept_hat: the estimated intercept

betahat_opt: the estimated beta with post-cellwise-robust regression

intercept_opt: the estimated intercept with post-cellwise-robust regression

...
}
\description{
robust sparse regression under cellwise contamination with a grid of lambdas (standardize predictors first)
}
\examples{
data = genevar()
y = data$y
x = data$x
fit = sregcell_std(y,x)


}
