\name{srlmcell-package}
\alias{srlmcell-package}
\alias{srlmcell}
\docType{package}
\title{
\packageTitle{srlmcell}
}
\description{
\packageDescription{srlmcell}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{srlmcell}
\packageIndices{srlmcell}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{srlmcell}

Maintainer: \packageMaintainer{srlmcell}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
