% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{genevar}
\alias{genevar}
\title{generate simulated datasets}
\usage{
genevar(
  n = 100,
  p = 10,
  pr = 5,
  e = 0,
  r = 0.5,
  df = Inf,
  beta = NULL,
  intercept = 0,
  sigma = 1,
  mux = rep(0, p),
  scalex = 1,
  gamma = 6,
  outtype = "cellwise"
)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of candidate predictors}

\item{pr}{number of active predictors}

\item{e}{contamination rate}

\item{r}{correlation among predictors}

\item{df}{degrees of freedom used to generate predictors}

\item{beta}{regression coefficients}

\item{intercept}{intercept of the regression model}

\item{sigma}{scale of the error item}

\item{mux}{averages of predictors}

\item{scalex}{scales of predictors}

\item{gamma}{magnitudes of outliers}

\item{outtype}{type of outliers: cellwise, rowwise}
}
\value{
x: the generated design matrix

xc: the generated design matrix without contamination

y: the generated response

ynew: another generated response (as an independent test set)

outlier: the generated outlier in the design matrix

erroroutlier: the generated outlier in response
}
\description{
generate simulated datasets
}
\examples{
data = genevar()
y = data$y
x = data$x

}
