% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcell-package.R
\docType{package}
\name{regcell-package}
\alias{regcell}
\alias{regcell-package}
\title{regcell: Cellwise regularized robust sparse regression}
\description{
Functions to perform sparse regression under cellwise contamination.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pengsu517.github.io/regcell/}
}

}
\author{
\strong{Maintainer}: Peng Su \email{peng.su@sydney.edu.au} (\href{https://orcid.org/0000-0002-1031-8675}{ORCID})

Authors:
\itemize{
  \item Garth Tarr \email{garth.tarr@gmail.com} (\href{https://orcid.org/0000-0002-6605-7478}{ORCID})
  \item Samuel Muller \email{samuel.muller@mq.edu.au} (\href{https://orcid.org/0000-0002-3087-8127}{ORCID})
  \item Suojin Wang \email{sjwang@stat.tamu.edu} (\href{https://orcid.org/0000-0001-6061-0052}{ORCID})
}

}
\keyword{internal}
