\name{regcell-package}
\alias{regcell-package}
\alias{regcell}
\docType{package}
\title{
\packageTitle{regcell}
}
\description{
\packageDescription{regcell}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{regcell}
\packageIndices{regcell}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{regcell}

Maintainer: \packageMaintainer{regcell}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
