% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcell.R
\docType{package}
\name{regcell}
\alias{regcell-package}
\alias{regcell}
\title{regcell}
\description{
Cellwise regularized robust sparse regression
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pengsu517.github.io/regcell/}
}

}
\author{
c(person("Peng", "Su", role = c("aut", "cre"), email = "peng.su@sydney.edu.au", comment = c(ORCID = "0000-0002-1031-8675")), person("Garth", "Tarr", role = "aut", email = "garth.tarr@gmail.com", comment = c(ORCID = "0000-0002-6605-7478")), person("Samuel", "Muller", role = "aut", email = "samuel.muller@mq.edu.au", comment = c(ORCID = "0000-0002-3087-8127")),person("Suojin", "Wang", role = "aut", email = "sjwang@stat.tamu.edu", comment = c(ORCID = "0000-0001-6061-0052")))
}
