% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcell-package.R
\name{regcell}
\alias{regcell}
\title{Cellwise regularized robust sparse regression}
\description{
Cellwise regularized robust sparse regression
}
\references{
Su P, Tarr G, Muller S and Wang S (2023). CR-Lasso: Robust cellwise regularized sparse regression. arXiv:2307.05234
}
\author{
c(person("Peng", "Su", role = c("aut", "cre"), email = "peng.su@sydney.edu.au", comment = c(ORCID = "0000-0002-1031-8675")), person("Garth", "Tarr", role = "aut", email = "garth.tarr@gmail.com", comment = c(ORCID = "0000-0002-6605-7478")), person("Samuel", "Muller", role = "aut", email = "samuel.muller@mq.edu.au", comment = c(ORCID = "0000-0002-3087-8127")),person("Suojin", "Wang", role = "aut", email = "sjwang@stat.tamu.edu", comment = c(ORCID = "0000-0001-6061-0052")))
}
\keyword{package}
